bin/publish: automatically deduce build tag if distserv tag is current
This allows one to publish the most recent build to the distribution server
by invoking 'publish -t current' (as is claimed by the current documentation
found in the comments on the top of bin/publish).

Note: when invoked in a build directory controlled by a CI system (e.g.,
buildbot), this is *NOT* a smart thing to do, as build/manifest.txt may be
changed by the build system (e.g., when the next build is started) w/o you
noticing.